feat: implement CSI simulator Go CLI for hardware-free testing
Implemented mothership/cmd/sim with full CSI simulation capabilities:
- Virtual nodes connect via WebSocket with hello/health/role protocol
- Synthetic CSI binary frame generation with proper header format
- Walker simulation with random walk motion and wall bouncing
- BLE advertisement simulation (optional)
- Blob count verification mode for CI integration
- CSV ground truth export for offline analysis
- Comprehensive test coverage

The simulator enables integration testing without ESP32 hardware:
  sim --mothership localhost:8080 --nodes 4 --walkers 2 --rate 20 --duration 60
  sim --verify --nodes 2 --walkers 1 --duration 10 --seed 42

+++ mothership/internal/replay/store.go
@@ -173,7 +173,7 @@ func (s *RecordingStore) Scan(fn func(recvTimeNS int64, frame []byte) bool) erro
 			return err
 		}
 		recvTimeNS := int64(binary.LittleEndian.Uint64(hdr[0:8]))
-		frameLen := int64(binary.LittleEndian.Uint64(hdr[8:10]))
+		frameLen := int64(binary.LittleEndian.Uint16(hdr[8:10]))
 		if frameLen > maxFrameBytes {
 			return errors.New("replay: corrupt record during scan")
 		}
@@ -222,7 +222,7 @@ func (s *RecordingStore) ScanRange(fromNS, toNS int64, fn func(recvTimeNS int64,
 			return err
 		}
 		recvTimeNS := int64(binary.LittleEndian.Uint64(hdr[0:8]))
-		frameLen := int64(binary.LittleEndian.Uint64(hdr[8:10]))
+		frameLen := int64(binary.LittleEndian.Uint16(hdr[8:10]))
 		if frameLen > maxFrameBytes {
 			return errors.New("replay: corrupt record during scan")
 		}
